# feat: Optimize pybind11 interface with zero-copy path handling 