--- a/mooncake-integration/store/store_py.cpp
+++ b/mooncake-integration/store/store_py.cpp
@@ -96,10 +96,7 @@ void ResourceTracker::signalHandler(int signal) {
     raise(signal);
 }
 
-void ResourceTracker::exitHandler() {
-    LOG(INFO) << "Process exiting, cleaning up resources";
-    getInstance().cleanupAllResources();
-}
+void ResourceTracker::exitHandler() { getInstance().cleanupAllResources(); }
 
 static bool isPortAvailable(int port) {
     int sock = socket(AF_INET, SOCK_STREAM, 0);
@@ -127,7 +124,6 @@ static int getRandomAvailablePort(int min_port = 12300, int max_port = 14300) {
 
     for (int attempts = 0; attempts < 10; attempts++) {
         int port = dis(gen);
-        std::cout << "port is " << port << std::endl;
         if (isPortAvailable(port)) {
             return port;
         }
@@ -431,12 +427,169 @@ int64_t DistributedObjectStore::getSize(const std::string &key) {
     return total_size;
 }
 
+// SliceBuffer implementation
+SliceBuffer::SliceBuffer(DistributedObjectStore &store,
+                         std::vector<mooncake::Slice> slices,
+                         uint64_t total_size)
+    : store_(store), slices_(std::move(slices)), total_size_(total_size) {}
+
+SliceBuffer::~SliceBuffer() {
+    if (consolidated_buffer_) {
+        delete[] consolidated_buffer_;
+        consolidated_buffer_ = nullptr;
+    }
+
+    // Free all slices
+    store_.freeSlices(slices_);
+}
+
+void *SliceBuffer::ptr() const {
+    if (slices_.empty()) {
+        return nullptr;
+    }
+    return slices_[0].ptr;
+}
+
+uint64_t SliceBuffer::size() const { return total_size_; }
+
+bool SliceBuffer::is_contiguous() const {
+    return slices_.size() == 1 || consolidated_buffer_ != nullptr;
+}
+
+void *SliceBuffer::consolidated_ptr() {
+    // If we already have a consolidated buffer, return it
+    if (consolidated_buffer_) {
+        return consolidated_buffer_;
+    }
+
+    // If there's only one slice, just return its pointer
+    if (slices_.size() == 1) {
+        return slices_[0].ptr;
+    }
+
+    // Otherwise, create a consolidated buffer
+    consolidated_buffer_ = new char[total_size_];
+    uint64_t offset = 0;
+    for (const auto &slice : slices_) {
+        memcpy(consolidated_buffer_ + offset, slice.ptr, slice.size);
+        offset += slice.size;
+    }
+
+    return consolidated_buffer_;
+}
+
+const std::vector<mooncake::Slice> &SliceBuffer::slices() const {
+    return slices_;
+}
+
+// Implementation of get_buffer method
+std::shared_ptr<SliceBuffer> DistributedObjectStore::get_buffer(
+    const std::string &key) {
+    if (!client_) {
+        LOG(ERROR) << "Client is not initialized";
+        return nullptr;
+    }
+
+    mooncake::Client::ObjectInfo object_info;
+    std::vector<Slice> slices;
+    uint64_t total_length = 0;
+    ErrorCode error_code;
+    std::shared_ptr<SliceBuffer> result = nullptr;
+
+    // Query the object info
+    error_code = client_->Query(key, object_info);
+    if (error_code != ErrorCode::OK) {
+        LOG(ERROR) << "Query failed for key: " << key
+                   << " with error: " << toString(error_code);
+        return nullptr;
+    }
+
+    // Allocate slices for the object
+    int ret = allocateSlices(slices, object_info, total_length);
+    if (ret) {
+        LOG(ERROR) << "Failed to allocate slices for key: " << key;
+        return nullptr;
+    }
+
+    // Get the object data
+    error_code = client_->Get(key, object_info, slices);
+    if (error_code != ErrorCode::OK) {
+        // Free slices on error
+        freeSlices(slices);
+        LOG(ERROR) << "Get failed for key: " << key
+                   << " with error: " << toString(error_code);
+        return nullptr;
+    }
+
+    // Create the SliceBuffer
+    result =
+        std::make_shared<SliceBuffer>(*this, std::move(slices), total_length);
+    return result;
+}
+
 PYBIND11_MODULE(store, m) {
+    // Define the SliceBuffer class
+    py::class_<SliceBuffer, std::shared_ptr<SliceBuffer>>(m, "SliceBuffer",
+                                                          py::buffer_protocol())
+        .def("ptr",
+             [](const SliceBuffer &self) {
+                 // Return the pointer as an integer for Python
+                 return reinterpret_cast<uintptr_t>(self.ptr());
+             })
+        .def("size", &SliceBuffer::size)
+        .def("is_contiguous", &SliceBuffer::is_contiguous)
+        .def_buffer([](SliceBuffer &self) -> py::buffer_info {
+            if (!self.is_contiguous()) {
+                void *consolidated_data = self.consolidated_ptr();
+                return py::buffer_info(
+                    consolidated_data, /* Pointer to buffer */
+                    sizeof(char),      /* Size of one scalar */
+                    py::format_descriptor<char>::
+                        format(), /* Python struct-style format descriptor */
+                    1,            /* Number of dimensions */
+                    {(size_t)self.size()}, /* Buffer dimensions */
+                    {sizeof(char)} /* Strides (in bytes) for each index */
+                );
+            } else if (self.size() > 0) {
+                return py::buffer_info(
+                    self.ptr(),   /* Pointer to buffer */
+                    sizeof(char), /* Size of one scalar */
+                    py::format_descriptor<char>::
+                        format(), /* Python struct-style format descriptor */
+                    1,            /* Number of dimensions */
+                    {(size_t)self.size()}, /* Buffer dimensions */
+                    {sizeof(char)} /* Strides (in bytes) for each index */
+                );
+            } else {
+                // Empty buffer
+                return py::buffer_info(
+                    nullptr,      /* Pointer to buffer */
+                    sizeof(char), /* Size of one scalar */
+                    py::format_descriptor<char>::
+                        format(),  /* Python struct-style format descriptor */
+                    1,             /* Number of dimensions */
+                    {0},           /* Buffer dimensions */
+                    {sizeof(char)} /* Strides (in bytes) for each index */
+                );
+            }
+        })
+        .def(
+            "consolidated_ptr",
+            [](SliceBuffer &self) {
+                // Return the consolidated pointer as an integer for Python
+                return reinterpret_cast<uintptr_t>(self.consolidated_ptr());
+            },
+            py::call_guard<py::gil_scoped_acquire>());
+
+    // Define the DistributedObjectStore class
     py::class_<DistributedObjectStore>(m, "MooncakeDistributedStore")
         .def(py::init<>())
         .def("setup", &DistributedObjectStore::setup)
         .def("init_all", &DistributedObjectStore::initAll)
-        .def("get", &DistributedObjectStore::get)  // Removed call_guard
+        .def("get", &DistributedObjectStore::get)
+        .def("get_buffer", &DistributedObjectStore::get_buffer,
+             py::call_guard<py::gil_scoped_release>(),
+             py::return_value_policy::take_ownership)
         .def("put", &DistributedObjectStore::put,
              py::call_guard<py::gil_scoped_release>())
         .def("remove", &DistributedObjectStore::remove,