--- a/mooncake-integration/store/store_py.h
+++ b/mooncake-integration/store/store_py.h
@@ -15,6 +15,7 @@ class DistributedObjectStore;
 
 // Forward declarations
 class SliceGuard;
+class SliceBuffer;
 
 // Global resource tracker to handle cleanup on abnormal termination
 class ResourceTracker {
@@ -49,10 +50,70 @@ class ResourceTracker {
     std::unordered_set<DistributedObjectStore *> instances_;
 };
 
+/**
+ * @brief A class that holds allocated slices and provides a contiguous view of
+ * the data This class is responsible for freeing the slices when it's destroyed
+ * (RAII)
+ */
+class SliceBuffer {
+   public:
+    /**
+     * @brief Construct a new SliceBuffer object
+     * @param store Reference to the DistributedObjectStore that owns the
+     * allocator
+     * @param slices Vector of slices that contain the data
+     * @param total_size Total size of the data across all slices
+     */
+    SliceBuffer(DistributedObjectStore &store,
+                std::vector<mooncake::Slice> slices, uint64_t total_size);
+
+    /**
+     * @brief Destructor that frees all slices
+     */
+    ~SliceBuffer();
+
+    /**
+     * @brief Get a pointer to the data
+     * @return void* Pointer to the data (first slice if multiple slices)
+     */
+    void *ptr() const;
+
+    /**
+     * @brief Get the size of the data
+     * @return uint64_t Size of the data in bytes
+     */
+    uint64_t size() const;
+
+    /**
+     * @brief Get a contiguous buffer containing all the data
+     * @return void* Pointer to the contiguous buffer (caller must NOT free
+     * this)
+     */
+    void *consolidated_ptr();
+
+    /**
+     * @brief Check if the data is stored in a single slice
+     * @return true if data is in a single slice, false otherwise
+     */
+    bool is_contiguous() const;
+
+    /**
+     * @brief Get the raw slices
+     * @return const std::vector<mooncake::Slice>& Reference to the slices
+     */
+    const std::vector<mooncake::Slice> &slices() const;
+
+   private:
+    DistributedObjectStore &store_;
+    std::vector<mooncake::Slice> slices_;
+    uint64_t total_size_;
+    char *consolidated_buffer_ = nullptr;
+};
+
 class DistributedObjectStore {
    public:
-    friend class SliceGuard;  // Allow SliceContainer to access private
-                              // members
+    friend class SliceGuard;   // Allow SliceGuard to access private members
+    friend class SliceBuffer;  // Allow SliceBuffer to access private members
     DistributedObjectStore();
     ~DistributedObjectStore();
 
@@ -70,6 +131,14 @@ class DistributedObjectStore {
     int put(const std::string &key, const std::string &value);
 
     pybind11::bytes get(const std::string &key);
+
+    /**
+     * @brief Get a buffer containing the data for a key
+     * @param key Key to get data for
+     * @return std::shared_ptr<SliceBuffer> Buffer containing the data, or
+     * nullptr if error
+     */
+    std::shared_ptr<SliceBuffer> get_buffer(const std::string &key);
 
     int remove(const std::string &key);