--- a/mooncake-integration/store/store_py.h
+++ b/mooncake-integration/store/store_py.h
@@ -128,7 +128,10 @@ class DistributedObjectStore {
     int initAll(const std::string &protocol, const std::string &device_name,
                 size_t mount_segment_size = 1024 * 1024 * 16);  // Default 16MB
 
-    int put(const std::string &key, const std::string &value);
+    int put(const std::string &key, std::span<const char> value);
+
+    int put_parts(const std::string &key,
+                  std::vector<std::span<const char>> values);
 
     pybind11::bytes get(const std::string &key);
 
@@ -166,6 +169,12 @@ class DistributedObjectStore {
     int allocateSlices(std::vector<mooncake::Slice> &slices,
                        const mooncake::Client::ObjectInfo &object_info,
                        uint64_t &length);
+
+    int allocateSlices(std::vector<mooncake::Slice> &slices,
+                       std::span<const char> value);
+
+    int allocateSlicesPacked(std::vector<mooncake::Slice> &slices,
+                             const std::vector<std::span<const char>> &parts);
 
     char *exportSlices(const std::vector<mooncake::Slice> &slices,
                        uint64_t length);