Directly return the internal buffer to the upper-level caller, instead of constructing a Python bytes object by copying.

performance result

### TCP
main

```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 1.00MB
Total data processed: 0.78GB
Put duration: 0.70 seconds
Get duration: 0.43 seconds
System Put throughput: 1149.84 ops/sec
System Get throughput: 1845.99 ops/sec
System Put bandwidth: 1.12 GB/sec
System Get bandwidth: 1.80 GB/sec
```

pr

```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 1.00MB
Total data processed: 0.78GB
Put duration: 0.71 seconds
Get duration: 0.29 seconds
System Put throughput: 1131.98 ops/sec
System Get throughput: 2728.87 ops/sec
System Put bandwidth: 1.11 GB/sec
System Get bandwidth: 2.66 GB/sec
```

### RDMA
main
```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 5.00MB
Total data processed: 3.91GB
Put duration: 1.32 seconds
Get duration: 2.59 seconds
System Put throughput: 605.09 ops/sec
System Get throughput: 308.60 ops/sec
System Put bandwidth: 2.95 GB/sec
System Get bandwidth: 1.51 GB/sec
```
pr
```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 5.00MB
Total data processed: 3.91GB
Put duration: 0.39 seconds
Get duration: 0.35 seconds
System Put throughput: 2062.64 ops/sec
System Get throughput: 2284.94 ops/sec
System Put bandwidth: 10.07 GB/sec
System Get bandwidth: 11.16 GB/sec
```